refactor(packages): one owner for the document notation's shared grammar
The .ttrs template format and the view engine's skeleton format are one
notation with two implementations. Per the 2026-08-07 one-document-format
decision — one notation, one parser — @transitrix/document-renderer, the
reference implementation of the directive language, now owns the grammar and
@transitrix/document-view-engine imports it instead of keeping a copy.

- document-renderer/src/ids.mjs absorbs isValidTypeName and becomes the single
  ID grammar; document-view-engine/src/ids.mjs (a byte-identical copy of its
  core, carrying a comment declaring the copy deliberate) is deleted.
- New document-renderer/src/syntax.mjs holds the syntax primitives both parsers
  carried identically: front matter, header scalars, header field reading, the
  identifier pattern, and the id/field-path split. Both parsers import it.
- Each parser keeps what is genuinely its own: its construct set (instruct here,
  each/trace/field-ref there), its error shape, and its AST node names. Those
  are implementation surface, not grammar, so parse-skeleton.mjs is delegated
  rather than retired — the smaller and more reversible of the two options.

Structural only, no behaviour change: all eight tests across both packages pass
unchanged, including the frozen conformance fixture's byte comparison.
